refactor: PrChangeStats.getTotalChanges() 반환 타입 int로 변경

+++ src/main/java/com/prism/statistics/domain/pullrequest/vo/PrChangeStats.java
@@ -56,7 +56,7 @@ public class PrChangeStats {
         this.deletionCount = deletionCount;
     }
 
-    public long getTotalChanges() {
-        return (long) additionCount + deletionCount;
+    public int getTotalChanges() {
+        return additionCount + deletionCount;
     }
 }
